+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 
 - Created basic GLUT window for visual debugging
-- Implemented simple vertex and fragment shader
-- `ShaderProgram` constructs a shader program from the filesystem
+- Implemented simple vertex and fragment shader for drawing a map
+- `DrawMapShader` supports rendering a `OccupancyGridTextureMap`
 
 [Unreleased]: https://github.com/jdelacroix/glamm/compare/v0.1.0...HEAD